backend/requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-cors==3.0.10
flask-jwt-extended==4.5.2
pymysql==1.0.3
cryptography==40.0.2
python-dotenv==1.0.0
gunicorn==20.1.0
insightface==0.7.3
onnxruntime==1.17.0
opencv-python-headless==4.7.0.72
numpy==1.24.3
Werkzeug==2.2.3
google-api-python-client==2.86.0
google-auth==2.17.3
Pillow==9.5.0
pandas==2.0.1
openpyxl==3.1.2